# Remove generics from IAvaloniaObject.