… src/Avalonia.Base/ApiCompatBaseline.txt
… src/Avalonia.Base/IAvaloniaObject.cs
… src/Avalonia.Base/StyledPropertyBase.cs
… src/Avalonia.Base/ValueStore.cs
… src/Avalonia.Controls/Primitives/AdornerLayer.cs
… src/Avalonia.Controls/Primitives/SelectingItemsControl.cs
… src/Avalonia.Controls/Repeater/ItemsRepeater.cs
… src/Avalonia.Controls/Repeater/RecyclePool.cs
… src/Avalonia.Controls/Repeater/ViewManager.cs
… src/Avalonia.Diagnostics/Diagnostics/ViewModels/ControlLayoutViewModel.cs
… src/Avalonia.Input/KeyboardNavigation.cs
… src/Avalonia.Input/Navigation/TabNavigation.cs
… src/Avalonia.Styling/StyledElement.cs
… src/Markup/Avalonia.Markup.Xaml.Loader/CompilerExtensions/Transformers/AvaloniaXamlIlWellKnownTypes.cs
… tests/Avalonia.Base.UnitTests/AvaloniaPropertyTests.cs
… tests/Avalonia.Markup.UnitTests/Data/BindingTests_TemplatedParent.cs
… tests/Avalonia.Styling.UnitTests/SetterTests.cs