## What does the pull request do?

We have a problem with our usage of generic virtual and generic interface methods, in that they cause performance problems in the JIT. We've decided that for 11.0 we need to at least remove generic interface methods from the public API. This PR removes generic methods from the `IAvaloniaObject` interface.

In order to do so, I had to throw exceptions in quite a few places if the `IAvaloniaObject` instance isn't an `AvaloniaObject`. Reimplementing this interface wasn't really feasible anyway so this shouldn't cause a problem in practise (and in future we will probably want to remove this interface completely but baby steps for now).

## Breaking Changes

Yes :P