@workgroupengineering 2022-04-12
For curiosity. How much does performance improve?