@grokys 2022-04-12
I doubt that it's changed much from just this PR - there are a couple of other more impactful changes that I'm about to make, and I'll benchmark once those are done. Wanted to separate the PRs though for easier review.