@grokys 2022-04-14
Full benchmarks after the first batch of changes at https://github.com/AvaloniaUI/Avalonia/pull/7980